docs : hooks file space utils
유틸 함수 파일 위치 변경 및 제거 작업
